chore: rename style file name
